+++ test/nextjs.test.js
@@ -13,17 +13,26 @@
  * under the License.
  */
 
-const { printPath, setupST, startST, killAllST, cleanST } = require("./utils");
+const { printPath, setupST, startST, killAllST, cleanST, delay } = require("./utils");
 let assert = require("assert");
 let { ProcessState } = require("../lib/build/processState");
 let SuperTokens = require("../lib/build/").default;
 let { middleware } = require("../framework/express");
 const Session = require("../lib/build/recipe/session");
 const EmailPassword = require("../lib/build/recipe/emailpassword");
+const EmailVerification = require("../lib/build/recipe/emailverification");
 const ThirdPartyEmailPassword = require("../lib/build/recipe/thirdpartyemailpassword");
-const superTokensNextWrapper = require("../lib/build/nextjs").superTokensNextWrapper;
+const {
+    superTokensNextWrapper,
+    withSession,
+    getSSRSession,
+    getAppDirRequestHandler,
+    withPreParsedRequestResponse,
+} = require("../lib/build/nextjs");
 const { verifySession } = require("../recipe/session/framework/express");
+const SessionError = require("../lib/build/recipe/session/error").default;
 const { testApiHandler } = require("next-test-api-route-handler");
+const { NextRequest, NextResponse } = require("next/server");
 
 let wrapperErr;
 
@@ -65,7 +74,7 @@ async function nextApiHandlerWithVerifySession(req, res) {
     }
 }
 
-describe(`NextJS Middleware Test: ${printPath("[test/nextjs.test.js]")}`, function () {
+describe(`Next.js Pages Router: ${printPath("[test/nextjs.test.js]")}`, function () {
     describe("with superTokensNextWrapper", function () {
         before(async function () {
             process.env.user = undefined;
@@ -136,7 +145,7 @@ describe(`NextJS Middleware Test: ${printPath("[test/nextjs.test.js]")}`, functi
                     assert.deepStrictEqual(respJson.user.emails[0], "john.doe@supertokens.io");
                     assert.strictEqual(respJson.user.id, process.env.user);
                     assert.notStrictEqual(res.headers.get("front-token"), undefined);
-                    const tokens = getSessionTokensFromResponse(res);
+                    const tokens = getSessionTokensFromResponseHeaders(res);
                     assert.notEqual(tokens.access, undefined);
                     assert.notEqual(tokens.refresh, undefined);
                 },
@@ -172,7 +181,7 @@ describe(`NextJS Middleware Test: ${printPath("[test/nextjs.test.js]")}`, functi
                     assert.deepStrictEqual(respJson.status, "OK");
                     assert.deepStrictEqual(respJson.user.emails[0], "john.doe@supertokens.io");
                     assert(res.headers.get("front-token") !== undefined);
-                    tokens = getSessionTokensFromResponse(res);
+                    tokens = getSessionTokensFromResponseHeaders(res);
                     assert.notEqual(tokens.access, undefined);
                     assert.notEqual(tokens.refresh, undefined);
                 },
@@ -606,9 +615,605 @@ describe(`NextJS Middleware Test: ${printPath("[test/nextjs.test.js]")}`, functi
     });
 });
 
-function getSessionTokensFromResponse(response) {
+describe(`Next.js App Router: ${printPath("[test/nextjs.test.js]")}`, function () {
+    before(async function () {
+        process.env.user = undefined;
+        await killAllST();
+        await setupST();
+        const connectionURI = await startST({ coreConfig: { access_token_validity: 2 } });
+        ProcessState.getInstance().reset();
+        SuperTokens.init({
+            supertokens: {
+                connectionURI,
+            },
+            appInfo: {
+                apiDomain: "api.supertokens.io",
+                appName: "SuperTokens",
+                apiBasePath: "/api/auth",
+                websiteDomain: "supertokens.io",
+            },
+            recipeList: [
+                EmailPassword.init(),
+                EmailVerification.init({}),
+                Session.init({
+                    override: {
+                        functions: (oI) => {
+                            return {
+                                ...oI,
+                                createNewSession: async (input) => {
+                                    let session = await oI.createNewSession(input);
+                                    process.env.user = session.getUserId();
+                                    return session;
+                                },
+                            };
+                        },
+                    },
+                }),
+            ],
+        });
+    });
+
+    after(async function () {
+        await killAllST();
+        await cleanST();
+    });
+
+    it("Sign Up", async function () {
+        const handleCall = getAppDirRequestHandler(NextResponse);
+
+        const userEmail = `john.doe.${Date.now()}@supertokens.com`;
+        const signUpRequest = new NextRequest("http://localhost:3000/api/auth/signup", {
+            method: "POST",
+            headers: { rid: "emailpassword" },
+            body: JSON.stringify({
+                formFields: [
+                    { id: "email", value: userEmail },
+                    { id: "password", value: "P@sSW0rd" },
+                ],
+            }),
+        });
+
+        const signUpRes = await handleCall(signUpRequest);
+        const respJson = await signUpRes.json();
+
+        assert.deepStrictEqual(respJson.status, "OK");
+        assert.deepStrictEqual(respJson.user.emails[0], userEmail);
+        assert.strictEqual(respJson.user.id, process.env.user);
+        assert.notStrictEqual(signUpRes.headers.get("front-token"), undefined);
+        const tokens = getSessionTokensFromResponseHeaders(signUpRes);
+        assert.notEqual(tokens.access, undefined);
+        assert.notEqual(tokens.refresh, undefined);
+    });
+
+    it("Sign In", async function () {
+        const handleCall = getAppDirRequestHandler(NextResponse);
+
+        const userEmail = `john.doe.${Date.now()}@supertokens.com`;
+        const requestInfo = {
+            method: "POST",
+            headers: { rid: "emailpassword" },
+            body: JSON.stringify({
+                formFields: [
+                    { id: "email", value: userEmail },
+                    { id: "password", value: "P@sSW0rd" },
+                ],
+            }),
+        };
+
+        const signUpRequest = new NextRequest("http://localhost:3000/api/auth/signup", requestInfo);
+
+        const signUpRes = await handleCall(signUpRequest);
+        assert.deepStrictEqual(signUpRes.status, 200);
+
+        const signInRequest = new NextRequest("http://localhost:3000/api/auth/signin", requestInfo);
+
+        const signInRes = await handleCall(signInRequest);
+        const respJson = await signInRes.json();
+
+        assert.deepStrictEqual(respJson.status, "OK");
+        assert.deepStrictEqual(respJson.user.emails[0], userEmail);
+        assert.strictEqual(respJson.user.id, process.env.user);
+        assert.notStrictEqual(signInRes.headers.get("front-token"), undefined);
+        const tokens = getSessionTokensFromResponseHeaders(signInRes);
+        assert.notEqual(tokens.access, undefined);
+        assert.notEqual(tokens.refresh, undefined);
+    });
+
+    it("getSSRSession", async function () {
+        const tokens = await getValidTokensAfterSignup({ tokenTransferMethod: "header" });
+
+        const authenticatedRequest = new NextRequest("http://localhost:3000/api/get-user", {
+            headers: {
+                Authorization: `Bearer ${tokens.access}`,
+            },
+        });
+
+        let sessionContainer = await getSSRSession(authenticatedRequest.cookies.getAll(), authenticatedRequest.headers);
+
+        assert.equal(sessionContainer.hasToken, true);
+        assert.equal(sessionContainer.session.getUserId(), process.env.user);
+
+        const unAuthenticatedRequest = new NextRequest("http://localhost:3000/api/get-user");
+
+        sessionContainer = await getSSRSession(unAuthenticatedRequest.cookies.getAll(), unAuthenticatedRequest.headers);
+
+        assert.equal(sessionContainer.hasToken, false);
+        assert.equal(sessionContainer.session, undefined);
+
+        const requestWithFailedClaim = new NextRequest("http://localhost:3000/api/get-user", {
+            headers: {
+                Authorization: `Bearer ${tokens.access}`,
+            },
+        });
+
+        sessionContainer = await getSSRSession(
+            requestWithFailedClaim.cookies.getAll(),
+            requestWithFailedClaim.headers,
+            {
+                overrideGlobalClaimValidators: async (globalValidators) => [
+                    ...globalValidators,
+                    EmailVerification.EmailVerificationClaim.validators.isVerified(),
+                ],
+            }
+        );
+        assert.equal(sessionContainer.hasToken, true);
+        assert.equal(sessionContainer.hasInvalidClaims, true);
+
+        await delay(3);
+        const requestWithExpiredToken = new NextRequest("http://localhost:3000/api/get-user", {
+            headers: {
+                Authorization: `Bearer ${tokens.access}`,
+            },
+        });
+
+        sessionContainer = await getSSRSession(
+            requestWithExpiredToken.cookies.getAll(),
+            requestWithExpiredToken.headers
+        );
+        assert.equal(sessionContainer.session, undefined);
+        assert.equal(sessionContainer.hasToken, true);
+    });
+
+    it("withSession", async function () {
+        const tokens = await getValidTokensAfterSignup({ tokenTransferMethod: "header" });
+
+        const authenticatedRequest = new NextRequest("http://localhost:3000/api/get-user", {
+            headers: {
+                Authorization: `Bearer ${tokens.access}`,
+            },
+        });
+
+        const authenticatedResponse = await withSession(authenticatedRequest, async (err, session) => {
+            if (err) return NextResponse.json(err, { status: 500 });
+            return NextResponse.json({
+                userId: session.getUserId(),
+                sessionHandle: session.getHandle(),
+                accessTokenPayload: session.getAccessTokenPayload(),
+            });
+        });
+
+        assert.equal(authenticatedResponse.status, 200);
+        const authenticatedResponseJson = await authenticatedResponse.json();
+        assert(authenticatedResponseJson.userId === process.env.user);
+
+        const requestWhereHandlerThrowsSTError = authenticatedRequest;
+        const responseWhereHandlerThrowsSTError = await withSession(
+            requestWhereHandlerThrowsSTError,
+            async (err, session) => {
+                throw new SessionError({ message: "Authentication Required!", type: "UNAUTHORISED" });
+            }
+        );
+        assert.equal(responseWhereHandlerThrowsSTError.status, 401);
+
+        const requestWhereHandlerThrowsUnknownError = authenticatedRequest;
+        const unknownError = new Error("Unknown error");
+        try {
+            await withSession(requestWhereHandlerThrowsUnknownError, async (err, session) => {
+                throw unknownError;
+            });
+            assert.fail("should not come here");
+        } catch (error) {
+            assert.strictEqual(error, unknownError);
+        }
+
+        const unAuthenticatedRequest = new NextRequest("http://localhost:3000/api/get-user");
+
+        const unAuthenticatedResponse = await withSession(unAuthenticatedRequest, async (err, session) => {
+            if (err) return NextResponse.json(err, { status: 500 });
+            return NextResponse.json({
+                userId: session.getUserId(),
+                sessionHandle: session.getHandle(),
+                accessTokenPayload: session.getAccessTokenPayload(),
+            });
+        });
+        assert.equal(unAuthenticatedResponse.status, 401);
+
+        const requestWithFailedClaim = new NextRequest("http://localhost:3000/api/get-user", {
+            headers: {
+                Authorization: `Bearer ${tokens.access}`,
+            },
+        });
+
+        const responseWithFailedClaim = await withSession(
+            requestWithFailedClaim,
+            async (err, session) => {
+                if (err) return NextResponse.json(err, { status: 500 });
+                return NextResponse.json({
+                    userId: session.getUserId(),
+                    sessionHandle: session.getHandle(),
+                    accessTokenPayload: session.getAccessTokenPayload(),
+                });
+            },
+            {
+                overrideGlobalClaimValidators: async (globalValidators) => [
+                    ...globalValidators,
+                    EmailVerification.EmailVerificationClaim.validators.isVerified(),
+                ],
+            }
+        );
+
+        assert.equal(responseWithFailedClaim.status, 403);
+
+        await delay(3);
+        const requestWithExpiredToken = new NextRequest("http://localhost:3000/api/get-user", {
+            headers: {
+                Authorization: `Bearer ${tokens.access}`,
+            },
+        });
+
+        const responseWithExpiredToken = await withSession(requestWithExpiredToken, async (err, session) => {
+            if (err) return NextResponse.json(err, { status: 500 });
+            return NextResponse.json({
+                userId: session.getUserId(),
+                sessionHandle: session.getHandle(),
+                accessTokenPayload: session.getAccessTokenPayload(),
+            });
+        });
+
+        assert.equal(responseWithExpiredToken.status, 401);
+
+        const requestThatThrows = {}; // this is an invalid request object and it will cause withSession to throw
+
+        const responseThatThrows = await withSession(requestThatThrows, async (err, session) => {
+            if (err) return NextResponse.json(err, { status: 500 });
+            throw new Error("test error");
+        });
+
+        assert.equal(responseThatThrows.status, 500);
+    });
+
+    it("withPreParsedRequestResponse", async function () {
+        const tokens = await getValidTokensAfterSignup({ tokenTransferMethod: "header" });
+
+        const authenticatedRequest = new NextRequest("http://localhost:3000/api/get-user", {
+            headers: {
+                Authorization: `Bearer ${tokens.access}`,
+            },
+        });
+
+        const authenticatedResponse = await withPreParsedRequestResponse(
+            authenticatedRequest,
+            async (baseRequest, baseResponse) => {
+                const session = await Session.getSession(baseRequest, baseResponse);
+
+                return NextResponse.json({ userId: session.getUserId() });
+            }
+        );
+
+        assert.equal(authenticatedResponse.status, 200);
+        const authenticatedResponseJson = await authenticatedResponse.json();
+        assert(authenticatedResponseJson.userId === process.env.user);
+
+        const requestWhereHandlerThrowsSTError = authenticatedRequest;
+        const responseWhereHandlerThrowsSTError = await withPreParsedRequestResponse(
+            requestWhereHandlerThrowsSTError,
+            async (baseRequest, baseResponse) => {
+                throw new SessionError({ message: "Authentication Required!", type: "UNAUTHORISED" });
+            }
+        );
+        assert.equal(responseWhereHandlerThrowsSTError.status, 401);
+
+        const requestWhereHandlerThrowsUnknownError = authenticatedRequest;
+        const unknownError = new Error("Unknown error");
+        try {
+            await withPreParsedRequestResponse(
+                requestWhereHandlerThrowsUnknownError,
+                async (baseRequest, baseResponse) => {
+                    throw unknownError;
+                }
+            );
+            assert.fail("should not come here");
+        } catch (error) {
+            assert.strictEqual(error, unknownError);
+        }
+    });
+});
+
+describe(`getSSRSession:hasToken`, function () {
+    describe("tokenTransferMethod = any", function () {
+        before(async function () {
+            process.env.user = undefined;
+            await killAllST();
+            await setupST();
+            const connectionURI = await startST();
+            ProcessState.getInstance().reset();
+            SuperTokens.init({
+                supertokens: {
+                    connectionURI,
+                },
+                appInfo: {
+                    apiDomain: "api.supertokens.io",
+                    appName: "SuperTokens",
+                    apiBasePath: "/api/auth",
+                    websiteDomain: "supertokens.io",
+                },
+                recipeList: [
+                    EmailPassword.init(),
+                    Session.init({
+                        getTokenTransferMethod: () => "any",
+                    }),
+                ],
+            });
+        });
+
+        after(async function () {
+            await killAllST();
+            await cleanST();
+        });
+
+        it("should return hasToken value correctly", async function () {
+            const tokens = await getValidTokensAfterSignup({ tokenTransferMethod: "header" });
+
+            const requestWithNoToken = new NextRequest("http://localhost:3000/api/get-user");
+
+            sessionContainer = await getSSRSession(requestWithNoToken.cookies.getAll(), requestWithNoToken.headers);
+
+            assert.equal(sessionContainer.hasToken, false);
+
+            const requestWithInvalidToken = new NextRequest("http://localhost:3000/api/get-user", {
+                headers: {
+                    Authorization: `Bearer some-random-token`,
+                },
+            });
+
+            sessionContainer = await getSSRSession(
+                requestWithInvalidToken.cookies.getAll(),
+                requestWithInvalidToken.headers
+            );
+            assert.equal(sessionContainer.hasToken, false);
+
+            const requestWithTokenInHeader = new NextRequest("http://localhost:3000/api/get-user", {
+                headers: {
+                    Authorization: `Bearer ${tokens.access}`,
+                },
+            });
+
+            sessionContainer = await getSSRSession(
+                requestWithTokenInHeader.cookies.getAll(),
+                requestWithTokenInHeader.headers
+            );
+            assert.equal(sessionContainer.hasToken, true);
+
+            const requestWithTokenInCookie = new NextRequest("http://localhost:3000/api/get-user", {
+                headers: {
+                    Cookie: `sAccessToken=${tokens.access}`,
+                },
+            });
+
+            sessionContainer = await getSSRSession(
+                requestWithTokenInCookie.cookies.getAll(),
+                requestWithTokenInCookie.headers
+            );
+            assert.equal(sessionContainer.hasToken, true);
+        });
+    });
+
+    describe("tokenTransferMethod = cookie", function () {
+        before(async function () {
+            process.env.user = undefined;
+            await killAllST();
+            await setupST();
+            const connectionURI = await startST();
+            ProcessState.getInstance().reset();
+            SuperTokens.init({
+                supertokens: {
+                    connectionURI,
+                },
+                appInfo: {
+                    apiDomain: "api.supertokens.io",
+                    appName: "SuperTokens",
+                    apiBasePath: "/api/auth",
+                    websiteDomain: "supertokens.io",
+                },
+                recipeList: [
+                    EmailPassword.init(),
+                    Session.init({
+                        getTokenTransferMethod: () => "cookie",
+                    }),
+                ],
+            });
+        });
+
+        after(async function () {
+            await killAllST();
+            await cleanST();
+        });
+
+        it("should return hasToken value correctly", async function () {
+            const tokens = await getValidTokensAfterSignup({ tokenTransferMethod: "cookie" });
+
+            const requestWithNoToken = new NextRequest("http://localhost:3000/api/get-user");
+
+            sessionContainer = await getSSRSession(requestWithNoToken.cookies.getAll(), requestWithNoToken.headers);
+
+            assert.equal(sessionContainer.hasToken, false);
+
+            const requestWithInvalidToken = new NextRequest("http://localhost:3000/api/get-user", {
+                headers: {
+                    Cookie: `sAccessToken=some-random-token`,
+                },
+            });
+
+            sessionContainer = await getSSRSession(
+                requestWithInvalidToken.cookies.getAll(),
+                requestWithInvalidToken.headers
+            );
+            assert.equal(sessionContainer.hasToken, false);
+
+            const requestWithTokenInHeader = new NextRequest("http://localhost:3000/api/get-user", {
+                headers: {
+                    Authorization: `Bearer ${tokens.access}`,
+                },
+            });
+
+            sessionContainer = await getSSRSession(
+                requestWithTokenInHeader.cookies.getAll(),
+                requestWithTokenInHeader.headers
+            );
+            assert.equal(sessionContainer.hasToken, false);
+
+            const requestWithTokenInCookie = new NextRequest("http://localhost:3000/api/get-user", {
+                headers: {
+                    Cookie: `sAccessToken=${tokens.access}`,
+                },
+            });
+
+            sessionContainer = await getSSRSession(
+                requestWithTokenInCookie.cookies.getAll(),
+                requestWithTokenInCookie.headers
+            );
+            assert.equal(sessionContainer.hasToken, true);
+        });
+    });
+
+    describe("tokenTransferMethod = header", function () {
+        before(async function () {
+            process.env.user = undefined;
+            await killAllST();
+            await setupST();
+            const connectionURI = await startST();
+            ProcessState.getInstance().reset();
+            SuperTokens.init({
+                supertokens: {
+                    connectionURI,
+                },
+                appInfo: {
+                    apiDomain: "api.supertokens.io",
+                    appName: "SuperTokens",
+                    apiBasePath: "/api/auth",
+                    websiteDomain: "supertokens.io",
+                },
+                recipeList: [
+                    EmailPassword.init(),
+                    Session.init({
+                        getTokenTransferMethod: () => "header",
+                    }),
+                ],
+            });
+        });
+
+        after(async function () {
+            await killAllST();
+            await cleanST();
+        });
+
+        it("should return hasToken value correctly", async function () {
+            const tokens = await getValidTokensAfterSignup({ tokenTransferMethod: "header" });
+
+            const requestWithNoToken = new NextRequest("http://localhost:3000/api/get-user");
+
+            sessionContainer = await getSSRSession(requestWithNoToken.cookies.getAll(), requestWithNoToken.headers);
+
+            assert.equal(sessionContainer.hasToken, false);
+
+            const requestWithInvalidToken = new NextRequest("http://localhost:3000/api/get-user", {
+                headers: {
+                    Authorization: `Bearer some-random-token`,
+                },
+            });
+
+            sessionContainer = await getSSRSession(
+                requestWithInvalidToken.cookies.getAll(),
+                requestWithInvalidToken.headers
+            );
+            assert.equal(sessionContainer.hasToken, false);
+
+            const requestWithTokenInHeader = new NextRequest("http://localhost:3000/api/get-user", {
+                headers: {
+                    Authorization: `Bearer ${tokens.access}`,
+                },
+            });
+
+            sessionContainer = await getSSRSession(
+                requestWithTokenInHeader.cookies.getAll(),
+                requestWithTokenInHeader.headers
+            );
+            assert.equal(sessionContainer.hasToken, true);
+
+            const requestWithTokenInCookie = new NextRequest("http://localhost:3000/api/get-user", {
+                headers: {
+                    Cookie: `sAccessToken=${tokens.access}`,
+                },
+            });
+
+            sessionContainer = await getSSRSession(
+                requestWithTokenInCookie.cookies.getAll(),
+                requestWithTokenInCookie.headers
+            );
+            assert.equal(sessionContainer.hasToken, false);
+        });
+    });
+});
+
+async function getValidTokensAfterSignup({ tokenTransferMethod = "header" } = {}) {
+    const handleCall = getAppDirRequestHandler(NextResponse);
+
+    const userEmail = `john.doe.${Date.now()}@supertokens.com`;
+    const requestInfo = {
+        method: "POST",
+        headers: { rid: "emailpassword" },
+        body: JSON.stringify({
+            formFields: [
+                { id: "email", value: userEmail },
+                { id: "password", value: "P@sSW0rd" },
+            ],
+        }),
+    };
+
+    const signUpRequest = new NextRequest("http://localhost:3000/api/auth/signup", requestInfo);
+
+    const signUpRes = await handleCall(signUpRequest);
+    assert.deepStrictEqual(signUpRes.status, 200);
+    const tokens =
+        tokenTransferMethod === "header"
+            ? getSessionTokensFromResponseHeaders(signUpRes)
+            : getSessionTokensFromResponseCookies(signUpRes);
+    assert.notEqual(tokens.access, undefined);
+    assert.notEqual(tokens.refresh, undefined);
+    return tokens;
+}
+
+function getSessionTokensFromResponseHeaders(response) {
     return {
         access: response.headers.get("st-access-token"),
         refresh: response.headers.get("st-refresh-token"),
     };
+}
+
+function getSessionTokensFromResponseCookies(response) {
+    const tokens = {};
+    response.headers.getSetCookie().forEach((header) => {
+        const matchAccessToken = header.match(/sAccessToken=([^;]+)/);
+        const matchRefreshToken = header.match(/sRefreshToken=([^;]+)/);
+
+        if (matchAccessToken) {
+            tokens.access = matchAccessToken[1];
+        }
+
+        if (matchRefreshToken) {
+            tokens.refresh = matchRefreshToken[1];
+        }
+    });
+
+    return tokens;
 }