+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [16.7.0] - 2023-12-22
+
+-   Add session util functions `withSession`, `getSSRSession` and `withPreParsedRequestResponse` for Next.js App directory.
+-   Previously, the custom framework's `errorHandler` callback function was invoked only upon encountering an error. This behavior has been rectified, and now the callback is invoked in both error and success cases.
+-   Added user creation apis on dashboard recipe to create Emailpassword and Passwordless recipe users.
+
 ## [16.6.8] - 2023-12-18
 
 -   Fix App dir support by removing the import of the "http" module